% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation_funs.R
\name{act_fun}
\alias{act_fun}
\alias{sigmoid}
\alias{ReLU}
\alias{ReU}
\alias{th}
\title{Activation Functions}
\usage{
sigmoid(x, zero = 1)

ReLU(x, zero = 0)

ReU(x, zero = 0)

th(x, zero = 0)
}
\arguments{
\item{x}{numeric input vector}

\item{zero}{numerical zero value for threshold}
}
\description{
Activation Functions
}
