% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.nn.R
\name{build.nn}
\alias{build.nn}
\title{Builds a Neural Network object}
\usage{
build.nn(inputs, outputs, hiddens = 0, weights = NULL,
  weight_range = c(-1, 1), enable_rate = 0.7)
}
\arguments{
\item{inputs, outputs}{Numeric indicating the number of inputs and outputs}

\item{hiddens}{Numeric Vector indicating the number of hidden neurons per layer (optional)}

\item{weights}{Numeric Vector of the nodal weight, if NULL (default) random weights
will be sampled from an uniform distribution over weight_range.
missing values will be randomized and excess values will be cropped}

\item{weight_range}{Numeric Vector of length 2 with the lower and
upper bounds for the random weight generation}

\item{enable_rate}{Numeric between 0 and 1 indicating the probability that a link will be enabled}
}
\value{
neural network object
}
\description{
Builds a Neural Network object
}
\examples{
build.nn(2, 1, enable_rate = 1) \%>\% plot
build.nn(2, 1, 2, enable_rate = 1) \%>\% plot
build.nn(2, 1, c(2,3), enable_rate = 1, weight_range = 0) \%>\% plot
build.nn(2, 2, c(5,5,5), enable_rate = 1, weight_range = 0) \%>\% plot
}
