% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{ntimestamp}
\alias{ntimestamp}
\title{neat representation of time stamp}
\usage{
ntimestamp(
  timestamp,
  include.date = TRUE,
  include.minutes = FALSE,
  include.seconds = FALSE,
  include.timezone = FALSE
)
}
\arguments{
\item{timestamp}{a POSIX time stamp}

\item{include.date}{a Boolean representing if the date of time stamp to be included}

\item{include.minutes}{a Boolean representing if the minutes to be included}

\item{include.seconds}{a Boolean representing if the seconds to be included}

\item{include.timezone}{a Boolean variable representing if the timezone of the date variable
to be included.}
}
\value{
String representation of time stamp
}
\description{
neat representation of time stamp
}
\examples{
x <- Sys.time()
ntimestamp(x)
}
