% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.nn.R
\name{populate.nn}
\alias{populate.nn}
\title{Build a population of Neural Nets}
\usage{
populate.nn(npop, inputs, outputs, hiddens = 0, weights = NULL,
  weight_range = c(-1, 1), enable_rate = 0.7)
}
\arguments{
\item{npop}{Numerical value of the number of individuals of the population}

\item{inputs}{Numeric indicating the number of inputs and outputs}

\item{outputs}{Numeric indicating the number of inputs and outputs}

\item{hiddens}{Numeric Vector indicating the number of hidden neurons per layer (optional)}

\item{weights}{Numeric Vector of the nodal weight, if NULL (default) random weights
will be sampled from an uniform distribution over weight_range.
missing values will be randomized and excess values will be cropped}

\item{weight_range}{Numeric Vector of length 2 with the lower and
upper bounds for the random weight generation}

\item{enable_rate}{Numeric between 0 and 1 indicating the probability that a link will be enabled}
}
\description{
This function generates a population of neural nets with the same input-output configuration
}
\examples{
populate.nn(10, 3, 1)
populate.nn(100, 3, 1)
}
